Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -2,8 +2,8 @@
 
 
 int main() {
-    float num;
-    scanf("%f", &num);
-    printf("You entered: %.4f", num);
+    double num;
+    scanf("%1f", &num);
+    printf("You entered: %1f", num);
     return 0;
 }
